vibrator: rework effects and fix bugs
Rework the EV_FF interfacing to avoid abusing the API, the new
implementation builds the effects during initialisation and uploads them
all, then when playing an effect we look up the id of the effect to play
and send the play command.

Then when off() is called we send a stop command for that effect ID.
This implementation lends itself much better to the EV_FF API and can
easily be expanded to support newer HAL versions with richer effects.

Sometimes seeing: "E VibratorService: Failed to issue command to
vibrator HAL. Retrying." but not sure what causes it.