audio: migrate back to yukawa HAL
This does the initial move to the newly "generic" yukawa audio HAL.

The HAL supports loading the mixer_paths.xml file based on some runtime
property (vendor.hw), this is used for dragonboard where the db845c and
rb5 use the same lunch target. However because we use different lunch
targets I've hardcoded vendor.hw to "sdm845", the mixer_paths.xml is
packaged from the device specific directory to
/vendor/etc/mixer_paths.sdm845.xml.

This depends on a new patch to system/media/ to support specifying the
output device in the mixer paths xml file:
https://android-review.googlesource.com/c/platform/system/media/+/2036767

I'll update our local_manifests to include this change in the mean time
before it's merged.

The beryllium and axolotl mixer paths are basically placeholder and will
need to be fixed. The enchilada paths *should* work, but for some reason
the speaker is broken, I suspect a kernel issue here. Headphone
detection and auto-switching works, on enchilada I'm able to hear audio
through headphones.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>