ril: UIM get card status works
woo!!! it's a complete mess right now, but it can successfully get the
state of the SIM card slots. A lot of cleanup is needed, and then work
on QMIC as right now the "auto generated" files need quite a bit of
poking.

Once the code is a bit cleaner it needs a proper way to handle
the asynchronicity of requests/responses, likely some multi-threading...
FUN!