lights: switch to AIDL HAL from somainline
The previously used android.hardware.light@2.0 is mostly just a
passthrough HAL that uses the legacy liblights that we implemented here.
The new HAL is using AIDL and implements the interface directly which is
much more what it should be.

This also fixes max brightness being (indirectly) hardcoded to 1020
instead of the actual values that devices support.

Link: https://github.com/SoMainline/android-lights-hal